# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Arize-Phoenix REST API",
    "description": "Schema for Arize-Phoenix REST API",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/experiments/{experiment_id}/runs": {
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "Create run for an experiment",
        "operationId": "createExperimentRun",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExperimentRunRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Experiment run created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateExperimentRunResponseBody"
                }
              }
            }
          },
          "403": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Experiment or dataset example not found"
          },
          "409": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "This experiment run has already been submitted"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "List runs for an experiment",
        "description": "Retrieve a paginated list of runs for an experiment",
        "operationId": "listExperimentRuns",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor for pagination (base64-encoded experiment run ID)",
              "title": "Cursor"
            },
            "description": "Cursor for pagination (base64-encoded experiment run ID)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "The max number of experiment runs to return at a time. If not specified, returns all results.",
              "title": "Limit"
            },
            "description": "The max number of experiment runs to return at a time. If not specified, returns all results."
          }
        ],
        "responses": {
          "200": {
            "description": "Experiment runs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListExperimentRunsResponseBody"
                }
              }
            }
          },
          "403": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Experiment not found"
          },
          "422": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Invalid cursor format"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateExperimentRunRequestBody": {
        "properties": {
          "dataset_example_id": {
            "type": "string",
            "title": "Dataset Example Id",
            "description": "The ID of the dataset example used in the experiment run"
          },
          "output": {
            "title": "Output",
            "description": "The output of the experiment task"
          },
          "repetition_number": {
            "type": "integer",
            "title": "Repetition Number",
            "description": "The repetition number of the experiment run"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time",
            "description": "The start time of the experiment run"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time",
            "description": "The end time of the experiment run"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id",
            "description": "The ID of the corresponding trace (if one exists)"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Optional error message if the experiment run encountered an error"
          }
        },
        "type": "object",
        "required": [
          "dataset_example_id",
          "output",
          "repetition_number",
          "start_time",
          "end_time"
        ],
        "title": "CreateExperimentRunRequestBody"
      },
      "CreateExperimentRunResponseBody": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateExperimentRunResponseBodyData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "CreateExperimentRunResponseBody"
      },
      "CreateExperimentRunResponseBodyData": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "The ID of the newly created experiment run"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "CreateExperimentRunResponseBodyData"
      },
...
      "ExperimentRunResponse": {
        "properties": {
          "dataset_example_id": {
            "type": "string",
            "title": "Dataset Example Id",
            "description": "The ID of the dataset example used in the experiment run"
          },
          "output": {
            "title": "Output",
            "description": "The output of the experiment task"
          },
          "repetition_number": {
            "type": "integer",
            "title": "Repetition Number",
            "description": "The repetition number of the experiment run"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time",
            "description": "The start time of the experiment run"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time",
            "description": "The end time of the experiment run"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id",
            "description": "The ID of the corresponding trace (if one exists)"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Optional error message if the experiment run encountered an error"
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "The ID of the experiment run"
          },
          "experiment_id": {
            "type": "string",
            "title": "Experiment Id",
            "description": "The ID of the experiment"
          }
        },
        "type": "object",
        "required": [
          "dataset_example_id",
          "output",
          "repetition_number",
          "start_time",
          "end_time",
          "id",
          "experiment_id"
        ],
        "title": "ExperimentRunResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListExperimentRunsResponseBody": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ExperimentRunResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "data",
          "next_cursor"
        ],
        "title": "ListExperimentRunsResponseBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}